add screenshot and adversarial tests
