+++ deploy/helm_charts/dc_website/templates/config_maps.yaml
@@ -69,4 +69,6 @@ metadata:
 data:
   webApiRoot: {{ .Values.cronTesting.webApiRoot | quote }}
   nodejsApiRoot: {{ .Values.cronTesting.nodejsApiRoot | quote }}
+  screenshotDomain: {{ .Values.cronTesting.screenshotDomain | quote }}
+  disableSanity: {{ .Values.cronTesting.disableSanity | quote }}
 {{- end }}
