+++ deploy/helm_charts/dc_website/values.yaml
@@ -145,3 +145,8 @@ cronTesting:
   # deployment runtime.
   webApiRoot:
   nodejsRoot:
+  # The domain to use for screenshot tests. Screenshot tests will only run if
+  # this is set.
+  screenshotDomain:
+  # To disable sanity tests, set this value to 1.
+  disableSanity: 0
