+++ deploy/helm_charts/envs/autopush.yaml
@@ -50,3 +50,4 @@ nodejs:
 
 cronTesting:
   enabled: true
+  screenshotDomain: "autopush.datacommons.org"
