+++ deploy/helm_charts/envs/bard.yaml
@@ -62,3 +62,4 @@ svg:
 
 cronTesting:
   enabled: true
+  disableSanity: 1
